feat: Refactor user and auth services, remove unused tests, and add new test cases for user registration and token verification
